replaced common GradingView with a separate ReleaseView to be used on the release page
